Merge new generator to trunk.
This is the first half of providing a stand-alone JSON class that can provide a
more flexible public interface than the category on various existing Foundation
classes.

This new interface returns NSError object via a pointer, if you request it.
It is also a bit faster at generating long strings than the previous code.


git-svn-id: https://json-framework.googlecode.com/svn/trunk@244 7ed384e2-a94a-0410-aa17-6fca9987d09a